move to components & refactor
