+++ packages/block-editor/README.md
@@ -654,22 +654,6 @@ _Returns_
 
 -   `Object`: Props to pass to the element to mark as a block.
 
-<a name="useCustomUnits" href="#useCustomUnits">#</a> **useCustomUnits**
-
-Custom hook to retrieve and consolidate units setting from add_theme_support().
-
-_Parameters_
-
--   _args_ `Object`: An object containing units, settingPath & defaultUnits.
--   _args.units_ `Object`: Collection of available units.
--   _args.settingPath_ `string`: The setting path. Defaults to 'spacing.units'.
--   _args.defaultUnits_ `Array`: Array of default units.
--   _args.defaultValues_ `Object`: Collection of default values for defined units. Example: { px: '350', em: '15' }.
-
-_Returns_
-
--   `Array`: Filtered units based on settings.
-
 <a name="useSetting" href="#useSetting">#</a> **useSetting**
 
 Hook that retrieves the editor setting.
