+++ packages/block-editor/src/components/index.js
@@ -76,10 +76,7 @@ export {
 	__unstableRichTextInputEvent,
 } from './rich-text';
 export { default as ToolSelector } from './tool-selector';
-export {
-	default as __experimentalUnitControl,
-	useCustomUnits,
-} from './unit-control';
+export { default as __experimentalUnitControl } from './unit-control';
 export { default as URLInput } from './url-input';
 export { default as URLInputButton } from './url-input/button';
 export { default as URLPopover } from './url-popover';
