+++ packages/block-library/src/cover/controls.native.js
@@ -17,11 +17,12 @@ import {
 	TextControl,
 	BottomSheet,
 	ToggleControl,
+	useCustomUnits,
 } from '@wordpress/components';
 import { plus } from '@wordpress/icons';
 import { useState, useCallback, useRef } from '@wordpress/element';
 import { usePreferredColorSchemeStyle } from '@wordpress/compose';
-import { useCustomUnits, MediaUpload } from '@wordpress/block-editor';
+import { useSetting, MediaUpload } from '@wordpress/block-editor';
 import { __ } from '@wordpress/i18n';
 
 /**
@@ -68,7 +69,13 @@ function Controls( {
 	);
 
 	const units = useCustomUnits( {
-		defaultUnits: [ 'px', 'em', 'rem', 'vw', 'vh' ],
+		availableUnits: useSetting( 'spacing.units' ) || [
+			'px',
+			'em',
+			'rem',
+			'vw',
+			'vh',
+		],
 		defaultValues: { px: '430', em: '20', rem: '20', vw: '20', vh: '50' },
 	} );
