+++ packages/components/src/unit-control/utils.js
@@ -7,6 +7,117 @@ import { isEmpty } from 'lodash';
  * WordPress dependencies
  */
 import { __, _x } from '@wordpress/i18n';
+import { Platform } from '@wordpress/element';
+
+const isWeb = Platform.OS === 'web';
+
+/**
+ * An array of all available CSS length units.
+ */
+export const ALL_CSS_UNITS = [
+	{
+		value: 'px',
+		label: isWeb ? 'px' : __( 'Pixels (px)' ),
+		default: '',
+		a11yLabel: __( 'Pixels (px)' ),
+	},
+	{
+		value: '%',
+		label: isWeb ? '%' : __( 'Percentage (%)' ),
+		default: '',
+		a11yLabel: __( 'Percent (%)' ),
+	},
+	{
+		value: 'em',
+		label: isWeb ? 'em' : __( 'Relative to parent font size (em)' ),
+		default: '',
+		a11yLabel: __( 'Relative to parent font size (em)' ),
+	},
+	{
+		value: 'rem',
+		label: isWeb ? 'rem' : __( 'Relative to root font size (rem)' ),
+		default: '',
+		a11yLabel: __( 'Relative to root font size (rem)' ),
+	},
+	{
+		value: 'vw',
+		label: isWeb ? 'vw' : __( 'Viewport width (vw)' ),
+		default: '',
+		a11yLabel: __( 'Viewport width (vw)' ),
+	},
+	{
+		value: 'vh',
+		label: isWeb ? 'vh' : __( 'Viewport height (vh)' ),
+		default: '',
+		a11yLabel: __( 'Viewport height (vh)' ),
+	},
+	{
+		value: 'vmin',
+		label: isWeb ? 'vmin' : __( 'Viewport smallest dimension (vmin)' ),
+		default: '',
+		a11yLabel: __( 'Viewport smallest dimension (vmin)' ),
+	},
+	{
+		value: 'vmax',
+		label: isWeb ? 'vmax' : __( 'Viewport largest dimension (vmax)' ),
+		default: '',
+		a11yLabel: __( 'Viewport largest dimension (vmax)' ),
+	},
+	{
+		value: 'ch',
+		label: isWeb ? 'ch' : __( 'Width of the zero (0) character (ch)' ),
+		default: '',
+		a11yLabel: __( 'Width of the zero (0) character (ch)' ),
+	},
+	{
+		value: 'ex',
+		label: isWeb ? 'ex' : __( 'x-height of the font (ex)' ),
+		default: '',
+		a11yLabel: __( 'x-height of the font (ex)' ),
+	},
+	{
+		value: 'lh',
+		label: isWeb ? 'lh' : __( 'Relative to the line-height (lh)' ),
+		default: '',
+		a11yLabel: __( 'Relative to the line-height (lh)' ),
+	},
+	{
+		value: 'cm',
+		label: isWeb ? 'cm' : __( 'Centimeters (cm)' ),
+		default: '',
+		a11yLabel: __( 'Centimeters (cm)' ),
+	},
+	{
+		value: 'mm',
+		label: isWeb ? 'mm' : __( 'Millimeters (mm)' ),
+		default: '',
+		a11yLabel: __( 'Millimeters (mm)' ),
+	},
+	{
+		value: 'Q',
+		label: isWeb ? 'Q' : __( 'Quarter-millimeters (Q)' ),
+		default: '',
+		a11yLabel: __( 'Quarter-millimeters (Q)' ),
+	},
+	{
+		value: 'in',
+		label: isWeb ? 'in' : __( 'Inches (in)' ),
+		default: '',
+		a11yLabel: __( 'Inches (in)' ),
+	},
+	{
+		value: 'pc',
+		label: isWeb ? 'pc' : __( 'Picas (pc)' ),
+		default: '',
+		a11yLabel: __( 'Picas (pc)' ),
+	},
+	{
+		value: 'pt',
+		label: isWeb ? 'pt' : __( 'Points (pt)' ),
+		default: '',
+		a11yLabel: __( 'Points (pt)' ),
+	},
+];
 
 /**
  * Units of measurements. `a11yLabel` is used by screenreaders.
@@ -77,7 +188,7 @@ export function hasUnits( units ) {
  * @param {Array<Object>} units Units to derive from.
  * @return {Array<number, string>} The extracted number and unit.
  */
-export function parseUnit( initialValue, units = CSS_UNITS ) {
+export function parseUnit( initialValue, units = ALL_CSS_UNITS ) {
 	const value = String( initialValue ).trim();
 
 	let num = parseFloat( value, 10 );
@@ -138,6 +249,48 @@ export function getValidParsedUnit( next, units, fallbackValue, fallbackUnit ) {
  * @return {string} a11y label for the unit abbreviation
  */
 export function parseA11yLabelForUnit( unit ) {
-	const match = CSS_UNITS.find( ( item ) => item.value === unit );
+	const match = ALL_CSS_UNITS.find( ( item ) => item.value === unit );
 	return match?.a11yLabel ? match?.a11yLabel : match?.value;
 }
+
+/**
+ * Filters available units based on values defined by settings.
+ *
+ * @param {Array} settings Collection of preferred units.
+ * @param {Array} units Collection of available units.
+ *
+ * @return {Array} Filtered units based on settings.
+ */
+function filterUnitsWithSettings( settings = [], units = [] ) {
+	return units.filter( ( unit ) => {
+		return settings.includes( unit.value );
+	} );
+}
+
+/**
+ * Custom hook to retrieve and consolidate units setting from add_theme_support().
+ *
+ * @param {Object} args                An object containing units, settingPath & defaultUnits.
+ * @param {Object} args.units          Collection of available units.
+ * @param {string} args.availableUnits The setting path. Defaults to 'spacing.units'.
+ * @param {Object} args.defaultValues  Collection of default values for defined units. Example: { px: '350', em: '15' }.
+ *
+ * @return {Array} Filtered units based on settings.
+ */
+export const useCustomUnits = ( { units, availableUnits, defaultValues } ) => {
+	units = units || ALL_CSS_UNITS;
+	const usedUnits = filterUnitsWithSettings(
+		! availableUnits ? [] : availableUnits,
+		units
+	);
+
+	if ( defaultValues ) {
+		usedUnits.forEach( ( unit, i ) => {
+			if ( defaultValues[ unit.value ] ) {
+				usedUnits[ i ].default = defaultValues[ unit.value ];
+			}
+		} );
+	}
+
+	return usedUnits.length === 0 ? false : usedUnits;
+};
